requirements.txt
matplotlib==3.1.2
imutils==0.5.3
opencv-python==4.1.2.30
Pillow==6.2.1
torch==1.3.0
torchvision==0.4.1
torchviz==0.0.1